+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
 ## Unreleased - XXXX-XX-XX
-- Migrate to bevy 0.18
-- Update flake
+- Technically breaking (But hidden behind plugin):
+  - Add `enable_debug` field to `TimeRunnerPlugin` by [#15](https://github.com/Multirious/bevy_time_runner/pull/15)
+  - Add `Tagging` variant to `TimeRunnerSet` by [#15](https://github.com/Multirious/bevy_time_runner/pull/15)
+  - Systems now expected `TimeCtx` generic parameter by [#15](https://github.com/Multirious/bevy_time_runner/pull/15)
+
+- Migrate to bevy 0.18 by [#16](https://github.com/Multirious/bevy_time_runner/pull/16)
+- Update flake by [#17](https://github.com/Multirious/bevy_time_runner/pull/17)
   - Use latest instead of a version for stableRust in flake.nix
   - `nix flake update`
   - Remove flake-utils dependency from flake.nix
+
+- Systems can now be registered in non-default time context. (Virtual, Fixed, and/or Real) See [#15](https://github.com/Multirious/bevy_time_runner/pull/15) for more details.
+
+  Notably:
+  - `TimeRunnerPlugin` which now uses `TimeRunnerSystemsPlugin<()>` by default.
+  - `TimeRunner` is always expected to have the `TimeContext<TimeCtx>` marker component.
+  - `TimeContext<TimeCtx>` is automatically inserted to children of `TimeRunner`.
+  - Add feature `debug`. This adds `TimeRunnerDebugPlugin` which logs warnings on missing `TimeContext<TimeCtx>` marker component when enabled.
 
 ## v0.5.2 - 2025-10-6
 - Fix documentation